Its funny, but I worked a lot on refactoring a tiny function
Basically, I tried to extract as much polymorphism as I can from a
lineList function.
I've came as close as I can at this moment, and not only the type now
looks gorgeous, it is *greatly* simplified!

The reason for this is that by using polymorphic types, there's much
more restriction on what a function can do, which means that the number
of possible implementations decreases.

This means less possible bugs!