Object is now F/A/M
In order to accomplish this, o_symbol is now a Symbol newtype that
instances Monoid on Char in the way that ' ' is treated as identity.

Also, added withCharacter and fromCharacter functions to help with the
fact that there are multiple state values going on, even if we know -
compile time - that the State must be Person.