Ability to store items in the containers and pull them out
No freaking idea why it works, to be honest.
Code is so fugly that it screams "REFACTOR ME!", but until I get clearer
understanding of what API should be - it'll stay the way it is.

The fault is all mine, tbh - instead of sitting down with paper and
pencil and actually architecting this out, I started coding and came up
with this.

Moron.