Code salvage from the obsolete `commands` branch
Rebase was impossibly difficult, and not much useful code was
written, so branch was nuked. Stuff that was kept:

* Direction and Iteration modules, which are genSingletons-ed
* Input system modifications to use Direction and Iteration
* Input GADTs use (Sing (d :: Direction)) as a proof for the phantom
  parameter. This means that withTyped*Event now correctly passes typed
  continuations
* Modification to Process* instances to work with new event kinds; this
  allows instances to be either specific or fully polymorphic
* Hud* instances now use closed type family to calculate returning
  GameState index

What's left to do is to see if I can congregate some instances together
using singleton's SEither. This would be useful to, for example, make an
instance that takes only (Move ('North || 'South)), and process only
those singletons in the functions.